+++ raven/visualizer/app.py
@@ -550,7 +550,11 @@ def _update_word_cloud(*, task_env):
                         keywords[kw] += count
 
                 logger.debug(f"_update_word_cloud: {task_env.task_name}: Invoking word cloud generator.")
-                wc = WordCloud(width=gui_config.word_cloud_w, height=gui_config.word_cloud_h, background_color="black", max_words=1000)
+                wc = WordCloud(width=gui_config.word_cloud_w,
+                               height=gui_config.word_cloud_h,
+                               background_color=gui_config.word_cloud_background_color,
+                               colormap=gui_config.word_cloud_colormap,
+                               max_words=1000)
                 wc.generate_from_frequencies(keywords)  # -> RGB tensor of shape [h, w, 3]
                 word_cloud_data_box << wc